# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Service API",
        "description": "API de gestion des utilisateurs",
        "contact": {
            "email": "n.taffot@elyft.tech"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/countries/code/{code}": {
            "get": {
                "tags": [
                    "Pays"
                ],
                "summary": "Afficher un pays par CODE",
                "description": "Retourne les informations d'un pays spécifique. Accessible uniquement au super_admin.",
                "operationId": "showCountryByCode",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "ID du pays à afficher",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "dkfjfoisofoikvfgvrr"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du pays",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "CM"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Cameroun"
                                        },
                                        "currency_code": {
                                            "type": "string",
                                            "example": "XAF"
                                        },
                                        "currency_symbol": {
                                            "type": "string",
                                            "example": "FCFA"
                                        },
                                        "is_active": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-07-15T14:25:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-07-15T14:25:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requête invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "PERMISSION_DENIED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pays introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pays introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/operators/code/{code}": {
            "get": {
                "tags": [
                    "Operators"
                ],
                "summary": "Voir un opérateur via son code",
                "operationId": "b7921de89539c8e4228770d641a25ae8",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'opérateur"
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Requête invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Accès refusé"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "PERMISSION_DENIED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ressource non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ressource non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}